requirements.txt
openai-whisper
yt-dlp
deep-translator
